Выделил логику задачи в отдельный презентер
