+++ src/presenter/presenter.js
@@ -5,6 +5,7 @@ import TripListView from '../view/trip-list-view';
 import NoPointsView from '../view/no-points-view';
 import { emptyListText } from '../view/no-points-view';
 import { render, RenderPosition } from '../framework/render';
+import PointPresenter from './point-presenter';
 
 export default class Presenter {
   constructor(
@@ -26,79 +27,41 @@ export default class Presenter {
 
     this.pointCardsModel = pointCardsModel;
 
-    this.tripListView = new TripListView(
-      (point) => this.tripListView.replacePointWithForm(point),
-      () => this.tripListView.replaceFormWithPoint(),
-      () => this.tripListView.replaceFormWithPoint()
-    );
+    this.tripListView = new TripListView();
+  }
 
-    this.rerenderPointsList = () => {
-      this.tripListView.removeElement();
-    };
+  /** Рендер информации в хедере */
+  renderHeaderTripInfo() {
+    render(
+      new HeaderTripInfoView(),
+      this.headerTripInfoContainer,
+      RenderPosition.AFTERBEGIN
+    );
+  }
+
+  /** Рендер фильтров Время */
+  renderHeaderTripTimeFilters() {
+    render(
+      this.headerTripTimeFiltersView,
+      this.headerTripControlsFilterContainer
+    );
+  }
+
+  /** Рендер фильтров в секции */
+  renderTripSectionFilters() {
+    render(new TripSectionListFilterView(), this.tripEventsSectionContainer);
+  }
+
+  /** Рендер точек */
+  renderPointCards() {
+    const pointPresenter = new PointPresenter();
+
+    pointPresenter.init(this.pointCardsModel.pointCards);
+
+    render(pointPresenter.pointCardContainer, this.tripEventsSectionContainer);
   }
 
   init() {
-    /** Загрузка фильтров Время */
-
-    // render(
-    //   this.headerTripTimeFiltersView,
-    //   this.headerTripControlsFilterContainer
-    // );
-
-    /** Заведение переменной для функции листенера кнопок Время */
-
-    // this.headerTripTimeFiltersViewListenerFunc = (evt) => {
-    //   this.noPointsView.removeElement();
-    //   this.noPointsView = new NoPointsView(emptyListText[evt.target.id]);
-    //   render(this.noPointsView, this.tripEventsSectionContainer);
-    // };
-
-    /** Заведение переменной для функции листенера кнопки Новая точка */
-
-    // this.headerAddNewPointButtonListenerFunc = () => {
-    //   this.headerTripTimeFiltersView.removeListener(
-    //     '',
-    //     'change',
-    //     this.headerTripTimeFiltersViewListenerFunc
-    //   );
-    //   this.noPointsView.removeElement();
-
-    //   this.headerAddNewPointButton.removeEventListener(
-    //     'click',
-    //     this.headerAddNewPointButtonListenerFunc
-    //   );
-
-    /** Рендер информации в хедере и списка ивентов */
-
-    //   render(
-    //     new HeaderTripInfoView(),
-    //     this.headerTripInfoContainer,
-    // RenderPosition.AFTERBEGIN
-    //   );
-
-    //   render(new TripSectionListFilterView(), this.tripEventsSectionContainer);
-    //   render(this.addNewPointTripListView, this.tripEventsSectionContainer);
-    // };
-
-    /** Рендер списка без точек */
-
-    // render(this.noPointsView, this.tripEventsSectionContainer);
-
-    /** Добавляем листенеры для фильтров Время и кнопки Новая точка */
-
-    // this.headerTripTimeFiltersView.addListener(
-    //   '',
-    //   'change',
-    //   this.headerTripTimeFiltersViewListenerFunc
-    // );
-
-    // this.headerAddNewPointButton.addEventListener(
-    //   'click',
-    //   this.headerAddNewPointButtonListenerFunc
-    // );
-
-    this.tripListView.init(this.pointCardsModel.pointCards);
-
-    render(this.tripListView, this.tripEventsSectionContainer);
+    this.renderPointCards();
   }
 }
